% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{smotif-package}
\alias{smotif}
\alias{smotif-package}
\title{smotif: Spatial Multi-Cellular Motifs}
\description{
Discovery and statistical evaluation of recurrent multi-cellular motifs (size 2, 3, and 4) in spatial omics data. Provides a lightweight 'SpatialGraph' representation backed by parquet, enumeration over igraph, spatially-aware null models, motif-to-gene association, and cross-region or cross-sample comparison. Designed to be framework-agnostic so that the core can be folded into Giotto Suite later.
}
\author{
\strong{Maintainer}: Ruben Dries \email{rdries@bu.edu}

}
\keyword{internal}
